make modular, created query
